Add chart of Proposed Blocks to Validator page
